Refurbished validationmessages
Number fields got "Required" insted of "Value is not a number" since the
input[type="number"] stopped the value being written to the model and we
had higher precedence on tv4 error messages. This is no longer the case
we show some ordinary angular validation messages as well.

This fixes #322

Also fixed some small bugs regarding the error messages. Often
`viewValue` is what you want in the error message.

Plus a start with protractor.