Unbreak validation
Validation was broken, mainly for numbers and date. This is an effort to do it in a more sane way.
It means we append our validation last, after angular has done transformation.

It has some side issues right now. Mainly typing charachters into a number
field that is also required will give you a "required" warning instead of
"this is not a number".