Removed optional bower dependencies
This is a tricky one. Schema Form has a lot of optional dependencies, basically it needs angular, angular-sanitize, tv4js and objectpath to work.
You probably need bootstrap 3 as well, but you might want to decide wich flavor so it's not specified. Angular 1.2 works but doesn't support
hyphens and other useful charachters as attribute names due to a bug in ng-model. This is fixed in angular 1.3, but then you loose IE8 support
(which is planned for schema form).

Check docs for what to include.